Add as_node_get_comment()
Comments are stored as internal attributes to avoid growing the size of
AsNodeData beyond 32 bits, although there is a performance penalty for this.
As comments are not often encountered in documents (and not at all in the
AppStream metadata fast path) this is probably okay.

By default comments are not loaded from the document, use
%AS_NODE_FROM_XML_FLAG_KEEP_COMMENTS when loading to enable this.