Add AS_STORE_ADD_FLAG_PREFER_LOCAL to allow the client to prefer local files
This allows an end application to use the local files rather than the AppStream
metadata.

This may be desired if testing the AppData file or viewing the screenshots.
It will certainly be slower as more files have to be parsed, more packages
resolved, and large screenshots will be downloaded just for thumbnails.

This allows a test mode: https://bugzilla.gnome.org/show_bug.cgi?id=732466