Save the screenshots as soon as possible
We don't want to store +4Gb of pixbuf data for the lifetime of the builder.
We can make the assuption that any application shipping screenshots also ships
a valid AppData file, and dump the screenshots to disk immediatly after the
task has completed. This means we can process Fedora rawhide with <2Gb RAM.