Close the packages after they have been processed
This means we have to re-open the files for the few packages used as
extra-packages to other packages, but means we keep the RSS value to a sane
value for a large number of packages. We can't go above 4GB RSS for Fedora...

This also means we spend less time 'shutting down' as the packages have
typically been released as part of the task, rather when the context exits.