Use the upstream screenshot when we are using appstream-builder --no-net
If we can't access the network we can't copy the screenshot onto a mirror or do
any of the resizing and padding that we normally do. This is also bad as there's
a slight privacy risk letting upstream know that the user is looking at some
specific software, and typically screenshots are not hosted on https:// so there
may be someone playing MITM games.

Also, but not resizing the screenshot we risk downloading the full size source
image just for the 112px thumbnail. But hey, it's better than nothing.

Resolves: https://github.com/hughsie/appstream-glib/issues/43