Do not expect the INF ClassGuid to be the ESRT GUID
I misunderstood the role of the ClassGuid for UEFI firmware. It is not supposed
to be the ESRT GUID, but instead a generic 'firmware' GUID set by Microsoft.
The actual ESRT is set as a FirmwareID registry key only, so learn how to parse
the HK section and fix up all the unit tests.

As there are .cab files in the wild (cough, ColorHug, cough) that set the
device ID as the ClassGuid, fall back to this to maintain compatibility with
either method.

Many thanks to Mario Limonciello for helping to debug this.

See https://msdn.microsoft.com/en-us/library/windows/hardware/ff547502%28v=vs.85%29.aspx
for more information.