Add as_release_set_filename()
This allows us to save the source package or bundle that is referenced to the
release. This allows us to download the correct file when installing a specific
version of a component.

This also allows us to mirror the correct file if the source filename is
different from the location-specified filename, or if there is no location
tag in the firmware metainfo file.