builder: Implement support for Solus .eopkg format
This format is employed by the Solus Operating System, and is implemented
in such a manner that it does not need to rely on executing the external
eopkg program, which is Pythonic, thus slower.

This utilises code that is part of an ongoing rewrite of eopkg, as such
in the future it would be refactored to use a 'libeopkg'. As such no pretense
is held to support the "PiSi" format, which the Pythonic eopkg is forked from,
as that support will not hold in the future in libeopkg.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>