Use a MS-style version number when parsing a hex version
If we're comparing versions that are both converted to strings it doesn't
really matter how the strings are stored (e.g. AA.BB.CC.DD or AA.BB.CCDD) as
the integer rules still apply **if** we don't remove the prefixed zeros.

The only time we have to be careful is where the firmware metainfo uses the
triplet and the .inf file uses the hex version format.

By using the triplet-style by default we match what Intel and Microsoft are
providing in the firmware release notes.

Basically:

 * If you're not using the MS-style versioning scheme, just use hex numbers in
   both places.

 * If you don't care about applying the update on Microsoft Windows, just
   don't supply an .inf file at all.

Fixes the other half of https://github.com/hughsie/fwupd/issues/34

This does break API, but we've never had the method in a released version so
I think it's fine to break at this stage. Make sure you update fwupd at the same
time if you're using git master.

The last bit of this fix would to be somehow encode the "display version" onto
the AppStream <release> object, but this needs discussion upstream first.