asb-package-ostree: Fix things up for xdg-app use
We only extract stuff from /export in order to not pick up
e.g. bundled desktop files not part of the app.

Also, we extract to a "usr" subdir instead of "export" to make
appstream-builder pick things up automatically (also remove the
previous hack that handled this by rewriting globs).