Move the blacklisted application list to the front-end client
At the moment the blacklist list is being used to filter applications that have
an AppData file, and also installed desktop files.

I agree it's wrong to filter applications with valid AppData files from the
metadata, and the various runtime blacklist entries are better suited to the
front-end application which can adjust it as the design dictates.