store: Pick up newly added appinfo dirs for xdg-app remotes
During the initial scan for appinfo dirs some appstream branches
may not have been updated yet. If this happens we don't create
any monitors for them, so we won't pick them up later.

So, we keep track of all xdg-app toplevel directories we monitor, and
we put a file monitor on them. Whenever these dirs change we rescan,
looking for new appinfo dirs.