Add an application prefix to the ID for certain install scopes
When we lived in a world where all software got installed to /usr we could
continue to use the application ID as the primary identifier.

Now we support installing things per-user, and also per-system and per-user
xdg-app (not even including jhbuild) we need to use the id prefix to
disambiguate the different applications according to a 'scope'.

This means when we launch a specific application in the software center
we know what desktop file to use, and we can also then support different
versions of applications installed system wide and per-user.

For the typical case of software installed in /usr by a package manager then
nothing changes, but if handling other kinds you need to audit the use of the
ID just in case you do want to share application data like reviews.