add missing "elementary" to list of project groups
Most official elementary software sets the \<project_group\> tag content to "elementary", which is not a recognised project group. This renders nearly all their .appdata.xml files invalid when trying to validate them with appstream-util.

This change would include "elementary" in the list of recognised project groups (I guessed this has to be added to the  table at libappstream-glib/as-app.c:4750). The "homepage" URL used is "http://elementary.io/".

I assumed the list is intentionally alphabetically sorted, so I inserted the missing line according to that.