Add ASB Extractor Plugin for Sugar Activities
Sugar Activities are a unique type of application desinged for Sugar.
Originally, they did not run outside of the Sugar DE.  However, they
now do and we will being shipping .desktop files for them [1].

We don't ship app data files.  It is much eaiser to just write a plugin
for ASB than write an app data file for every activity.  Having a
ASB plugin is also beneficial for us as it allows us to tag clearly
all Sugar Activities as being Sugar Activities.  This would allow us
to use App Stream as a basis for a Sugar Activity store and updater.

Additioanlly, it allows GNOME Software to have richer metadata when
accessing the Sugar Activities.  Sugar Activities already have
summaries, translations and icons, which are beneficial for
GNOME Software.

[1]  https://github.com/sugarlabs/sugar-toolkit-gtk3/pull/321