Allow setting the id prefix and origin using a symlink file name
The idea is to encode the scope and user-specified origin in the filename as
flatpak remotes have metadata with the filename `appstream.xml.gz` that have
origin harcoded to 'flatpak'.

We can't set a custom origin when generating the metadata as the remote name
is set client side and can be any string the user chooses.

It's also expected that something like gnome-software will keep the symlink
cache up to date. I'm using ~/.local/share/app-info/xmls at the moment, but if
we decide that the data is actually gnome-software specific then it can move.