as-app-desktop: Fix association between icon names and types
AsIcon documents that STOCK icons should have the name set; and that
LOCAL icons should have the filename set. The previous approach to
extracting these properties from as_app_parse_file_key() could sometimes
result in a LOCAL icon with no filename set, if tmp was not an icon name
and not a fallback icon name. Now, we assume that (as long as tmp is not
an absolute path) this means it's an icon name for an icon which is not
installed, and hence the AsIcon type should be STOCK.

Signed-off-by: Philip Withnall <philip.withnall@collabora.co.uk>