Pay attention to errors from yaml_parser_parse
The previous commit message blames libyaml for spinning forever on
invalid input.  This is not the case.  libyaml is correctly returning
an error code, but as_node_yaml_process_layer was ignoring it and
spinning forever calling back into libyaml as if expecting the error
to disappear by itself.

The code should probably be audited for other ignored error codes, but
this fixes the immediate problem.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>