Remove the inf parsing from appstream-glib
I don't want to maintain this large chunk of complex code in the main library,
especially as it's basically unused except for the validation of .cab files.
The latter has a python implmentation as used on the LVFS, so it's really just
dead code now.

Move the (unused) implementation to libginf for archival, but leave stubs in to
avoid upsetting the API/ABI checkers. We can remove the symbols when we next
break soname.