Process file changes when an attribute changes
When a file's permissions are changed this may mean that a file that was
previously unreadable can now be parsed so we should process files attribute
changes.

Also, when a file was changed, no processing of it occurred until a different
event type happened to the file (deletion, renaming, etc.). This could prevent
important changes to be processed. Update the tests for both issues.

Signed-off-by: Richard Hughes <richard@hughsie.com>