Detect invalid files in the libyaml read handler
If non-gzipped files are saved as .yml.gz they are seeked by the GInputStream
GZlibDecompressor which fails.

This can happen when the user is connected to a badly configured captive portal
and the downloaded 'gzipped yaml file' is actually the captive portal HTML
login page. Detect this and handle the error the best we can given the libyaml
API constraints.

This probably fixes bugs like https://github.com/hughsie/fwupd/issues/70
and similar bugs in gnome-software.