Recursively add subpackage deps into the main package
This fixes appstream data generation for packages that have long dep
chains and icons are at a deeper level than in the immediate child
package. e.g. in Fedora, 'marble' package depends on 'marble-widget-qt5'
which depends on 'marble-astro' which depends on 'marble-common' where
we can finally find the app's icon.