AsStore: hold a ref in hash_unique_id to avoid use-after-free
If we rely on the reference here being "borrowed" from priv->array,
then it's easy for that assumption to become unjustified. This
caused a crash for me in as_store_remove_app_by_id, which removes
the app from priv->array before it removes it from
priv->hash_unique_id; if no other code is holding a ref to the same
app, then as_app_get_unique_id() will be a use-after-free.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>