Add a destructor for AsRefString
AsRefString uses a singleton hash table for deduplicating strings. This
is all "possibly lost" memory according to valgrind as the hash table
never gets destroyed.

This commit adds a custom destructor so that we can clean up the
singleton hash table as well and avoid unnecessary noise in valgrind
output.