Deprecate app_set_source_file() and app_set_source_kind()
Implement these in terms of AsFormat objects to maintain API compatibility.

This allows us to have multiple 'formats' that make up an AsApp, for instance,
an application might be referenced in:

 * The .desktop file (for the icon)
 * The AppData file (for the installed long description)
 * The AppStream XML file (for the origin information)

Doing this allows us to keep track of all formats that make up an AsApp,
and this allows us to update the state as each type is removed and added.

This allows us to remove the need to "promote" apps to different types, which
was unreliable and was the source of several bugs where the AsApp was removed
from the AsStore before all the sources had been removed.

With this commit, GNOME Software again works reliably.