Copy hash table keys from AsApp to avoid them being used after they've been freed
The hash tables currently reference the keys in the AsApp object, however this
fails in two cases:

1. If you insert two apps with the same ID into the same table the second app
replaces the first but the key is from the original one (confirmed in glib
source code). The first app is unreffed leaving the table to contain an invalid
key (or may occur later if a reference is held elsewhere).

2. There are two tables that contain GPtrArrays of apps. If AsApp in the array
that was used as the key is unreffed then the hash table again contains a freed
key.

I'm assuming these keys weren't copied to save memory, if that is important then
the keys will have to be reference counted or atoms used.