meson: avoid unnecessary gdk dependency
gdk is only needed by the asb_plugin_font, which only gets built when
both appstream-build and fonts are enabled. Therefore checking for the
dependency should also be limited to that case, otherwise gdk has to
be provided during compilation also in cases where it is not used at
all.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>