Write XML for newer AppStream specification versions
Any XML parser should ignore the extra tags or attributes, and so it serves no
purpose to artificailly hide these when building metadata for older versions of
the specification.

This also fixes a bug on i386 where 0.6 != 0.6 and I didn't want to use fabs()
all over the place for basically no reason.