Fix the arithmetic when fitting an image in 16:9
When saving a 16:9 pixbuf and the image parameter is not 16:9 we create
a transparent pixbuf with this ratio and fit the image in it. However,
this calculation was not correctly done due to an integer division so
the result is that the width/height of the pixbuf's area to be copied
was bigger than what the dimensions of the pixbuf; so no image was
copied and we'd be left with a trasnparent pixbuf.

This patch fixes this problem by using multiplications instead of
a division in the mentioned code (as they avoid imprecision problems).