Correctly validate files using OR in the metadata_license
Althought the correct way to do this would be to construct a GTree structure,
splitting on AND and brackets, and having children of OR, I don't have a
spare weekend to implement this and then write all the tests.

This simplistic parser will correctly validate more files and that's good
enough for now.

Fixes https://github.com/hughsie/appstream-glib/issues/226