as-image: Handle images which are smaller than the destination
We currently don't consider this case at all, which results in negative
coordinates being passed to gdk_pixbuf_copy_area().

Let's instead look and see if the source image is smaller than the
destination. If it is, we just need to put it in the middle of our
desintation pixbuf.