Add AS_APP_QUIRK_DEVELOPER_VERIFIED
This quirk allows apps to marked as having an additional level of checking
performed on the developer. This is to give confidence to the user that this
is the official version of an app. This is a common design pattern used in
modern packaging systems (e.g. Apple App Store, Google Play).