client: handle ctrl-c on Windows with SetConsoleCtrlHandler().
g_unix_signal_add_full() (as all of glib-unix.h API) is not available on
Windows, as expected from the name.

It seems that signal() could be used, though apparently
SetConsoleCtrlHandler() may be the most appropriate method to handle a
ctrl-c, according to Microsoft docs.

The code is still untested but at least it builds.