as-content-rating: Add missing OARS → CSM mappings to fix round trips
Currently, it’s not possible to do a round trip from a CSM age to a
set of OARS ratings and back to a CSM age without the age changing; this
is because there are no mappings for ages 16 and 17. Fix that by adding
them.

There is no child psychology knowledge behind the choices of ages here:
they were chosen purely to fit in with the existing ratings and to fix
round trips.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://github.com/hughsie/appstream-glib/issues/280