Fix GUID creation from string/data on Windows.
There was endianness issues on Windows. We were just storing the data in
array of bytes on the Linux implementation, but the Win32 type UUID is
storing part of the identifier as long, other parts as short, and
finally the rest as chars. This messes up the byte order.
So we make sure (with htonl and htons) that we memcpy() data converted
from whatever host order to network byte order to always be consistent.