Set XDG_DATA_DIRS before loading the unit-test store.
The unit tests are assuming that the XDG base directory spec is in
usage. Under Windows, this is not the case by default, and in particular
g_get_system_data_dirs() would return the Application Data and Documents
folders for all users.
This can be overrided if XDG_DATA_DIRS is explicitly set. Since a
destdir is also set, it will look for store data in the source folder.