build: Properly check for the rpm option
Currently, building appstream-glib with `-Dbuilder=false -Drpm=true`
fails:

https://gnome1.codethink.co.uk/logs/build-2019-02-24-091101/build-gnome-apps-nightly-master-org.gnome.Software-aarch64.txt

This is because in the top-level meson.build file, the `rpm` variable
is only defined when both the `builder` and `rpm` options are true.

That means in libappstream-glib/meson.build we cannot just check for the
`rpm` option and assume the variable was defined.