as-content-rating: Add content rating system APIs from gnome-software
These are moving to libappstream-glib as they are used by gnome-software
and libmalcontent-ui, and are potentially useful in other projects. In
particular, there are a number of translatable strings which need to be
centralised so that translation work isn’t repeated.

This code was originally written by:
 • Richard Hughes <richard@hughsie.com>
 • Piotr Drąg <piotrdrag@gmail.com>
 • Philip Withnall <withnall@endlessm.com>
 • Jordi Mas <jmas@softcatala.org>

It was previously in `src/gs-content-rating.[ch]` in gnome-software,
under the GPL-2.0+; and in `libmalcontent-ui/gs-content-rating.[ch]` in
malcontent, also under the GPL-2.0+. It has been relicensed from
GPL-2.0+ to LGPL-2.1+ for the move to appstream-glib, and all the
previous contributors have agreed. See
https://github.com/hughsie/appstream-glib/pull/364#issuecomment-632773905
and below for their agreements.

In moving it from one project to the other, I have added documentation
comments and tweaked some of the APIs to make them more usable.

Signed-off-by: Philip Withnall <withnall@endlessm.com>