libappstream-builder: Fix file paths in icon tarball
dbd62f6e05 removed the WxH directory from the icon filenames, which were also
used as-is for the paths in the icons tarball. However, if width and height
are set on the icon, then it has to be inside a size-specific directory.
With HiDPI enabled, it did even overwrite the non-HiDPI icon it saved just
before because the filenames were the same.
Add the subdirectory back in the resource handling code.