Fix extracting relative symlinks
ebdefa2745 tried to fix handling of relative symlinks in archives, but that
only addressed targets starting with ".."/. Fix it properly by using
asb_utils_sanitise_path only for absolute paths, keep relative paths as-is.

Simplify code for handling absolute symlinks by just prepending the root
instead of converting it to a relative link.

Hardlink targets are returned as paths relative to the archive root, so the
sanitization code is correct there.